+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 😺 [Kitten Scraper ](https://github.com/skylarstein/kitten-scraper)Change Log
 
+## [1.8.0](https://github.com/skylarstein/kitten-scraper/compare/v1.7.1...v1.8.0) (2018-11-11)
+
+* Retrieve current animal status, filter out animals not currently in foster
+* Cleanup output CSV, add current animal status, remove unnecessary data, sort
+
 ## [1.7.1](https://github.com/skylarstein/kitten-scraper/compare/v1.7.0...v1.7.1) (2018-11-11)
 * Improved debug output and error handling
 * Fixes Python3 support